requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT==0.2.0
Flask-JWT-Extended==3.18
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
PyJWT==1.7.1
pytz==2021.1
six==1.16.0
SQLAlchemy==1.3.9
Werkzeug==2.0.1
pytest==6.2.5
gunicorn==20.1.0
apache-airflow==2.2.0
python-dotenv==0.19.1